+++ plugins/chakra/src/stories/example.stories.tsx
@@ -1,0 +1,9 @@
+import type { Meta } from '@storybook/react'
+
+const meta: Meta = {
+  title: 'Examples',
+}
+
+export { AuthenticationCard } from '~/examples/authentication-card'
+
+export default meta