+++ plugins/chakra/src/theme/components/heading.ts
@@ -1,7 +1,14 @@
-export const Heading = {
-  baseStyle: {
+import { defineStyle, defineStyleConfig } from '@chakra-ui/react'
+
+export const Heading = defineStyleConfig({
+  baseStyle: defineStyle((props) => ({
     color: 'fg.default',
     fontWeight: 'semibold',
     textStyle: 'xl',
+  })),
+  defaultProps: {
+    colorScheme: 'accent',
   },
-}
+  sizes: {},
+  variants: {},
+})