+++ plugins/chakra/src/theme/index.ts
@@ -6,36 +6,57 @@ import { config } from './config'
 import { direction } from './direction'
 import { fontSizes } from './font-sizes'
 import { fontWeights } from './font-weights'
+import { fonts } from './fonts'
 import { letterSpacings } from './letter-spacing'
 import { lineHeights } from './line-heights'
-import { radii } from './radii'
+import { aliases, radii } from './radii'
 import { semanticTokens } from './semantic-tokens'
 import { sizes } from './sizes'
 import { space } from './space'
 import { styles } from './styles'
 import { textStyles } from './text-styles'
 import { transition } from './transition'
+import type { Options } from './types'
 import { zIndices } from './z-indices'
 
-export const theme = {
-  borders,
-  breakpoints,
-  colors,
-  components,
-  config,
-  direction,
-  fonts: {},
-  fontSizes,
-  fontWeights,
-  letterSpacings,
-  lineHeights,
-  radii,
-  semanticTokens,
-  shadows: {},
-  sizes,
-  space,
-  styles,
-  textStyles,
-  transition,
-  zIndices,
+const defaultOptions: Options = {
+  accentColor: 'neutral',
+  grayColor: 'neutral',
+  borderRadius: 'sm',
+}
+
+export const createTheme = (options: Options = defaultOptions) => {
+  const { accentColor, grayColor, borderRadius = 'sm' } = options
+  return {
+    borders,
+    breakpoints,
+    colors,
+    components,
+    config,
+    direction,
+    fonts,
+    fontSizes,
+    fontWeights,
+    letterSpacings,
+    lineHeights,
+    radii: {
+      ...radii,
+      ...aliases[borderRadius],
+    },
+    semanticTokens: {
+      ...semanticTokens,
+      colors: {
+        ...semanticTokens.colors,
+        gray: semanticTokens.colors[grayColor],
+        accent: semanticTokens.colors[accentColor],
+      },
+    },
+    shadows: {},
+    sizes,
+    space,
+    styles,
+    textStyles,
+    transition,
+    zIndices,
+  }
 }