+++ plugins/chakra/src/theme/radii.ts
@@ -9,9 +9,42 @@ export const radii = {
   '2xl': '1rem',
   '3xl': '1.5rem',
   full: '9999px',
+}
 
-  // aliases
-  l1: '0.125rem',
-  l2: '0.25rem',
-  l3: '0.375rem',
+export const aliases = {
+  none: {
+    l1: '0',
+    l2: '0',
+    l3: '0',
+  },
+  xs: {
+    l1: '0.0625rem',
+    l2: '0.125rem',
+    l3: '0.25rem',
+  },
+  sm: {
+    l1: '0.125rem',
+    l2: '0.25rem',
+    l3: '0.375rem',
+  },
+  md: {
+    l1: '0.25rem',
+    l2: '0.375rem',
+    l3: '0.5rem',
+  },
+  lg: {
+    l1: '0.375rem',
+    l2: '0.5rem',
+    l3: '0.75rem',
+  },
+  xl: {
+    l1: '0.5rem',
+    l2: '0.75rem',
+    l3: '1rem',
+  },
+  '2xl': {
+    l1: '0.75rem',
+    l2: '1rem',
+    l3: '1.5rem',
+  },
 }