+++ plugins/panda/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+### Changed
+
+- Renamed `placement` prop to `variant` in `Drawer` component for better consistency with other CSS frameworks.
+- Renamed `Label` to `FormLabel`
+
 ## [0.27.0] - 2023-12-19
 
 ### Added